% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scca_core.R
\name{align_omics}
\alias{align_omics}
\title{Align Multi-Omic Datasets}
\usage{
align_omics(X, Y)
}
\arguments{
\item{X}{Matrix or data frame (Samples x Features).}

\item{Y}{Matrix or data frame (Samples x Features).}
}
\value{
A list containing the aligned X and Y matrices.
}
\description{
Ensures that X and Y matrices have matching samples in the exact same order.
}
