% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_plotting.R
\name{plot_correlation_heatmap}
\alias{plot_correlation_heatmap}
\title{Global Gene-Metabolite Correlation Heatmap}
\usage{
plot_correlation_heatmap(scca_model, X, Y, top_n = 20)
}
\arguments{
\item{scca_model}{The result object from omic_scca().}

\item{X}{Aligned RNA-seq matrix.}

\item{Y}{Aligned Metabolomics matrix.}

\item{top_n}{Number of top features from each omic to include.}
}
\value{
(Invisible) A numeric matrix of correlations between the selected features in \code{X} and \code{Y}.
}
\description{
Visualizes the correlation matrix with a gradient color scale.
}
