% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_plotting.R
\name{plot_pathway_circle}
\alias{plot_pathway_circle}
\title{Canonical Loading Pathway Circle Plot}
\usage{
plot_pathway_circle(scca_model, top_features = 40, pathway_db = "KEGG")
}
\arguments{
\item{scca_model}{The result object from omic_scca().}

\item{top_features}{Number of most weighted features to map.}

\item{pathway_db}{Conceptual database name for labeling.}
}
\value{
A ggplot2 object.
}
\description{
Visualizes top feature importance in a radial layout.
}
